requirements.txt
django==3.0.5
bcrypt==3.1.5=
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0